refactor(ff-filter): split graph.rs and filter_inner.rs into subdirectories
